backend/back/requirements.txt
aiofiles==23.2.1
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
exceptiongroup==1.2.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
priority==2.0.0
python-dotenv==1.0.1
python-nmap==0.7.1
Quart==0.19.6
quart-auth==0.10.1
taskgroup==0.0.0a4
tomli==2.0.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2